Fuel is now float

+++ app/src/main/java/edu/ub/pis2016/pis16/strikecom/gameplay/items/Item.java
@@ -11,12 +11,12 @@ public abstract class Item implements Comparable<Item> {
 	protected String image;    // item icon
 	protected String model;   // name of representative image on the screen
 	protected String flavour; //description, fluff
-	protected float price;    // price in game resources
+	protected int price;    // price in game resources
 
 	protected boolean visible = true;
 
 	// Builders
-	public Item(String name, String image, String model, String flavour, float price) {
+	public Item(String name, String image, String model, String flavour, int price) {
 		this.name = name;
 		this.image = image;
 		this.model = model;
@@ -36,7 +36,7 @@ public abstract class Item implements Comparable<Item> {
 
 	public String getFlavour(){ return this.flavour; }
 
-	public float getPrice(){ return this.price; }
+	public int getPrice(){ return this.price; }
 
 	// Returns negative integer if object price is smaller than that of a given object, positive
 	// integer if price is bigger and zero if both have equal price.
