+++ app/src/main/java/edu/ub/pis2016/pis16/strikecom/gameplay/items/UpgradeItem.java
@@ -17,7 +17,7 @@ public class UpgradeItem extends Item{
 
 	private String function;
 
-	public UpgradeItem(String name, String image, String model, String flavour, String function, float price) {
+	public UpgradeItem(String name, String image, String model, String flavour, String function, int price) {
 		super(name, image, model, flavour, price);
 		this.function = function;
 	}
@@ -35,7 +35,7 @@ public class UpgradeItem extends Item{
 		if(param.length < 6){ //should at least contain name, image, model, flavour, function, price
 			return null;
 		}
-		return new UpgradeItem(param[0], param[1], param[2], param[3], param[4], Float.valueOf(param[5]));
+		return new UpgradeItem(param[0], param[1], param[2], param[3], param[4], Integer.valueOf(param[5]));
 	}
 
 	public static boolean functionIsSupported(String f) {
